Code cleanup

+++ src/Migrations/20200129194920-create-general-miner-info.js
@@ -16,7 +16,7 @@ module.exports = {
                 type: sequelize.BIGINT
             },
             numberOfSectors: {
-                type: sequelize.BIGINT
+                type: sequelize.INTEGER
             },
             minerPower: {
                 type: sequelize.BIGINT
